종료 시간이 시작 시간과 같을 때 종료 시간을 1초 앞당겨 표현하던 문제 수수정
